+++ src/index.ts
@@ -11,6 +11,7 @@ import Common from "./common/index";
 import { WbiSign } from "./base/sign";
 import { fakeBuvid3, fakeDmCoverImgStr } from "./utils";
 import { protobufDecode, protoBufToXml } from "./utils/dm";
+import Downloader from "./utils/downloader";
 
 declare module "axios" {
   export interface AxiosRequestConfig {
@@ -122,4 +123,5 @@ export {
   Platform,
   Live,
   Auth,
+  Downloader,
 };